Add a 2nd rhythm and use model/view/update pattern
This starts the app down the path of the Elm Architecture. Although
there are no actions implemented yet, the view now renders the correct
rhythm based on the model (which has 2 possible choices now).

Some links are also rendered with fake hrefs as the beginning of a
selection UI for which rhythm is displayed.